requirements.txt
numpy==1.23.3
pandas==1.4.3
python-binance==1.0.16
python-dotenv==0.20.0
matplotlib==3.5.3
mplfinance==0.12.9b1
python-telegram-bot==21.1
proplot==0.9.5
requests~=2.28.1
pytest==7.1.2
pytest-mock==3.12.0
freezegun==1.2.2
tqdm==4.64.1
pybit==5.6.2
joblib==1.2.0
lightgbm==4.0.0
scikit-learn==1.2.2
